Some updates for #162

+++ core/addr.py
@@ -9,7 +9,7 @@ import re
 
 def addr_to_int(value):
     _ = value.split('.')
-    return (long(_[0]) << 24) + (long(_[1]) << 16) + (long(_[2]) << 8) + long(_[3])
+    return (int(_[0]) << 24) + (int(_[1]) << 16) + (int(_[2]) << 8) + int(_[3])
 
 def int_to_addr(value):
     return '.'.join(str(value >> n & 0xff) for n in (24, 16, 8, 0))
